Created Notes on Gaussian Quadrature
A brief overview that provides motivation for technique.

TODO discuss error, and implement it in MatLab, choose some functions
and do some testing.

Functions with singularities/oscillate quickly/discontinuities, are
probably bad so we should include them, theres a rocket equation
everyone seems to use in examples so include that asw